Refactored EditableLogic
